pyproject.toml
[project]
name = "weather-mcp-server"
version = "1.0.0"
description = "FastMCP Weather Server with OpenWeather API integration"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "real_api: marks tests that make real API calls (deselect with '-m \"not real_api\"')",
]
